+++ common/utilities/os/CMakeLists.txt
@@ -1,8 +1,9 @@
 # License: Apache 2.0. See LICENSE file in root directory.
-# Copyright(c) 2019 Intel Corporation. All Rights Reserved.
+# Copyright(c) 2021 Intel Corporation. All Rights Reserved.
 
-target_sources(${LRS_TARGET}
-    PRIVATE
-        "${CMAKE_CURRENT_LIST_DIR}/hresult.h"
+if(WIN32)
+    target_sources(${LRS_TARGET}
+        PRIVATE
+            "${CMAKE_CURRENT_LIST_DIR}/hresult.h")
 
-)
+endif()
