+++ common/utilities/string/string-utilities.h
@@ -8,6 +8,10 @@
 #include <sstream>
 #include <cmath> // std::isfinite
 
+#ifdef WIN32
+#include <Windows.h>
+#endif
+
 namespace utilities {
 namespace string {
     inline std::string hexify(unsigned char n)
@@ -189,6 +193,31 @@ namespace string {
             return false;
         }
     }
+
+#ifdef WIN32
+    inline std::string win_to_utf(const WCHAR * s)
+    {
+        auto len = WideCharToMultiByte(CP_UTF8, 0, s, -1, nullptr, 0, nullptr, nullptr);
+        std::stringstream ss;
+
+        if (len == 0)
+        {
+            ss << "WideCharToMultiByte(...) returned 0 and GetLastError() is " << GetLastError();
+            throw std::runtime_error(ss.str());
+        }
+
+        std::string buffer(len - 1, ' ');
+        len = WideCharToMultiByte(CP_UTF8, 0, s, -1, &buffer[0], static_cast<int>(buffer.size()) + 1, nullptr, nullptr);
+        if (len == 0)
+        {
+            ss.clear();
+            ss << "WideCharToMultiByte(...) returned 0 and GetLastError() is " << GetLastError();
+            throw std::runtime_error(ss.str());
+        }
+
+        return buffer;
+    }
+#endif
 
 }  // namespace string
 }  // namespace utilities
