+++ src/mf/mf-uvc.cpp
@@ -35,7 +35,7 @@ The library will be compiled without the metadata support!\n")
 #include <vidcap.h>
 #include <ksmedia.h>    // Metadata Extension
 #include <Mferror.h>
-#include "win/win-helpers.h"
+#include "../common/utilities/string/string-utilities.h"
 
 #pragma comment(lib, "Shlwapi.lib")
 #pragma comment(lib, "mf.lib")
@@ -55,7 +55,7 @@ namespace librealsense
 {
     namespace platform
     {
-        using namespace utilities::hresult;
+        using namespace utilities::string;
 
 #ifdef METADATA_SUPPORT
 
@@ -269,7 +269,7 @@ namespace librealsense
                 reinterpret_cast<void **>(&ks_topology_info)));
 
             DWORD nNodes=0;
-            CHECK_HR_STR("get_NumNodes", ks_topology_info->get_NumNodes(&nNodes));
+            CHECK_HR_STR("get_NumNodes", ks_topology_info->get_NumNodes(&nNodes), false);
 
             CComPtr<IUnknown> unknown = nullptr;
             CHECK_HR(ks_topology_info->CreateNodeInstance(xu.node, IID_IUnknown,
@@ -893,7 +893,7 @@ namespace librealsense
                     {
                         safe_release(pMediaType);
                         if (hr != MF_E_NO_MORE_TYPES) // An object ran out of media types to suggest therefore the requested chain of streaming objects cannot be completed
-                            check("_reader->GetNativeMediaType(sIndex, k, &pMediaType.p)", hr, false);
+                            LOG_HR(hr);
 
                         break;
                     }
@@ -1005,7 +1005,7 @@ namespace librealsense
                                 const auto timeout_ms = RS2_DEFAULT_TIMEOUT;
                                 if (_has_started.wait(timeout_ms))
                                 {
-                                    CHECK_HR_STR("_reader->ReadSample(...)", _readsample_result);
+                                    CHECK_HR_STR("_reader->ReadSample(...)", _readsample_result, false);
                                 }
                                 else
                                 {
