+++ src/win/win-helpers.h
@@ -16,18 +16,6 @@
 
 #define WAIT_FOR_MUTEX_TIME_OUT  (5000)
 
-#define CHECK_HR_STR(s, x) utilities::hresult::check(s, x, [](std::string descr, bool to_throw)\
-{\
-    if (to_throw)\
-        throw librealsense::windows_backend_exception(descr);\
-    else\
-        LOG_DEBUG(descr);\
-});
-
-#define CHECK_HR(x) CHECK_HR_STR(#x, x)
-
-#define LOG_HR(x) utilities::hresult::check(#x, x, false)
-
 namespace librealsense
 {
     namespace platform
